Fixed hero targeting
This commit makes it so that the spells automatically target whichever
is closest to the actual mouse click. If the hero is in the middle, it
casts the hero spell. If a bad guy is closest to the middle, it casts
on all baddies in range. Also, fixed the hero spells to time-out
correctly. It resets the speed and damage ratio now, once the effect
has worn off. Ideally, we'd want to highlight which GameObject the
spell is targeting, so that the user knows which one they're targeting.