Draws a rudimentary indicator over the current target.
Basically, this just draws a GUI box over either the hero or the bad
guy that's closest to the center of the target. We can change this to
be an icon of some sort, or possibly change the color, or whatever we
want to do. This is just a start.